refactor: move engine to services.process
