+++ src/aap_eda/services/process/engine/common.py
@@ -20,7 +20,7 @@ from datetime import datetime
 from django.conf import settings
 from pydantic import BaseModel, validator
 
-import aap_eda.services.activation.engine.exceptions as exceptions
+import aap_eda.services.process.engine.exceptions as exceptions
 from aap_eda.core.enums import ActivationStatus
